+++ src/EntityGraphQL/Directives/DirectiveLocation.cs
@@ -1,0 +1,13 @@
+namespace EntityGraphQL.Directives;
+
+public enum ExecutableDirectiveLocation
+{
+    QUERY,
+    MUTATION,
+    SUBSCRIPTION,
+    FIELD,
+    FRAGMENT_DEFINITION,
+    FRAGMENT_SPREAD,
+    INLINE_FRAGMENT,
+    VARIABLE_DEFINITION,
+}